Separate out tessellation ray tracing logic
